# vioscsi: set SCSISTAT_QUEUE_FULL if a buffer cannot be added to a VQ